refactor: move security configuration to the shared package
